invois/requirements.txt
appdirs==1.4.4
asgihandler==0.3.9
asgiref==3.6.0
astroid==2.3.3
async-timeout==4.0.2
black==23.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.1
colorama==0.4.3
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.9
django-admin-rangefilter==0.9.0
django-bootstrap3==12.1.0
django-ckeditor==5.9.0
django-cors-headers==3.13.0
django-crispy-forms==1.8.1
django-debug-toolbar==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-leaflet==0.28.3
django-login-required-middleware==0.4
django-redis==5.2.0
django-taggit==1.2.0
django-timezone-field==4.0
django-user-visit==0.4
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-csv==2.1.1
docopt==0.6.2
et-xmlfile==1.1.0
ez-setup==0.9
Faker==4.1.0
filelock==3.11.0
flake8==3.8.4
geojson==3.0.1
gunicorn==20.0.4
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-metadata==5.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nes-py==8.2.1
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pathspec==0.11.1
pbr==5.11.1
Pillow==9.4.0
pipenv==2023.3.20
pipreqs==0.4.12
platformdirs==3.1.0
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycryptodome==3.17
pyflakes==2.2.0
pyglet==1.5.21
pylint==2.4.4
pyparsing==3.0.9
python-dateutil==2.8.1
pytz==2020.1
redis==4.3.4
requests==2.28.1
six==1.14.0
sqlparse==0.3.1
stevedore==5.0.0
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2022.5
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.26.12
user-agents==2.2.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.10.0
